% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{bqte}
\alias{bqte}
\title{Estimate backtransformed quantile treatment effects}
\usage{
bqte(
  Treatment,
  Control,
  at = NULL,
  bqte.conf = 0.95,
  B = 2000,
  K = length(Control),
  bagging = TRUE,
  tails = FALSE,
  discrete.range = FALSE,
  exact.sample.quantiles = TRUE,
  interpolation.method = "linear",
  spline.df = 5,
  verbose = TRUE
)
}
\arguments{
\item{Treatment}{vector of outcome values in Treatment group.}

\item{Control}{vector of outcome values in Control group.}

\item{at}{vector of outcome values in controls at which BQTE is estimated.
By default, between 10 and 20 points between the empirical quantile levels
of 10/N and 1 - (10/N) in controls where N = min(|Treatment|, |Control|).}

\item{bqte.conf}{(default 0.95), confidence level for BQTE.}

\item{B}{(default 2000), number of bootstrap samples.}

\item{K}{(default 'length(Control)), number of quantiles where BQTE function is estimated.}

\item{bagging}{(default TRUE), if TRUE, reports bagging estimate (mean over bootstrap samples)
otherwise reports the direct estimate from the data.}

\item{tails}{(default FALSE), if TRUE, returns also estimates and confidence intervals for
UTBQTE and LTBQTE.}

\item{discrete.range}{(default FALSE), if TRUE, returns, for every Control group value,
the range of Treatment group values that correspond
through the shared quantile levels.}

\item{exact.sample.quantiles}{(default TRUE), if TRUE, uses empirical sample quantiles,
that are called type = 1 in R's quantile function,
if FALSE, uses R's default quantiles (type = 7).}

\item{interpolation.method}{either "linear" for piecewise-linear model or "spline"
for natural splines of df 'spline.df'.}

\item{spline.df}{(default 5) degrees of freedom of natural spline.
Relevant only if interpolation.method = "spline".}

\item{verbose}{(default TRUE) if TRUE, prints parameters and range recommendations on console.}
}
\value{
list with 1,2, or 3 data frames

$res

'at', outcome value in control group;

'bqte', backtransformed quantile treatment effect (averaged in case of ties);

'bqte.low', lower CI end point for 'bqte';

'bqte.up', upper CI end point for 'bqte';

'rbqte', relative backtransformed quantile treatment effect with respect to control outcome value;

'rbqte.low', lower CI end point for 'rbqte';

'rbqte.up', upper CI end point for 'rbqte';

$discrete.range, returned only if argument discrete.range = TRUE;

'at', outcome value in control group;

'range.low', lower CI end point of raw, non-averaged BQTE values corresponding to this Control group value;

'range.up' upper CI end point;

'rrange.low' lower CI end point of relative BQTE;

'rrange.up' upper CI end point of relative BQTE;

$tails, returned only if parameter 'tails = TRUE'

'utbqte, upper tail average treatment effect estimate;

'utbqte.low, lower CI end point for 'utbqte';

'utbqte.up, upper CI end point for 'utbqte';

'ltbqte, lower tail average treatment effect estimate;

'ltbqte.low, lower CI end point for 'ltbqte';

'ltbqte.up, upper CI end point for 'ltbqte';

'rutbqte, relative upper tail average treatment effect estimate with respect to control value;

'rutbqte.low, lower CI end point for 'rutbqte';

'rutbqte.up, upper CI end point for 'rutbqte';

'rltbqte, relative lower tail average treatment effect estimate with respect to control value;

'rltbqte.low, lower CI end point for 'rltbqte';

'rltbqte.up, upper CI end point for 'rltbqte'
}
\description{
We have observed survival/recovery times, or other numerical outcome values,
in the treatment group and in the control group.
Quantile treatment effect (QTE) is the difference between
the two groups as a function of quantile level in (0,1).
We want to transform QTE to the original scale of the outcome value,
at a given grid of values ('at') in the control group.
These transformed values are called
backtransformed quantile treatment effects' (BQTE).
}
\details{
Approach:
Fix a grid of outcome values in the control group at which
BQTEs are computed (parameter 'at').
Recommendation:
  Do not try to estimate BQTEs at values that are outside
  empirical quantile levels (10/N, 1 - (10/N)),
  where N is the minimum sample size of the treatment and control groups
  because accuracy of quantile estimates is low at the tails.

Estimate the quantile treatment effects as difference between the quantiles
of Treatment group and Control group at 'K' quantile levels
1/(K+1),...,K/(K+1).
Make a piece-wise linear approximation of BQTE
as a function of outcome value in controls,
based on the observed K quantiles.
If several quantiles are equal, then use average of the corresponding
Treatment group values as the value of BQTE at that point.
(See parameter 'discrete.range' to get more details of range of tied values.)
Use that approximation
to estimate the BQTE at the chosen grid of outcome values in controls.
NOTE: values are returned to the user on the grid defined by 'at',
      with a sensible default range, and not all K values included in the computation
      are meant to be returned to the user.

Uses bootstrapping to estimate the uncertainty of BQTE at each grid point,
by resampling with replacement a set of Treatment and Control group values that
have the same sample size as the original Treatment and Control groups, respectively,
and applying the procedure explained above to each bootstrap sample.
Note that bootstrapping accounts for uncertainty of quantiles of both groups.

The final estimate of BQTE at a grid point is the mean over bootstrap samples ("bagging estimate")
and the confidence interval for BQTE is estimated from the quantiles of the bootstrap sample.
NOTE: In empirical evaluation it was observed that bagging estimate tend to have smaller
      mean square error in discrete distributions than the direct estimate from the data and
      therefore the bagging estimate is used by default.
      This can be switched by argument "bagging".

For discrete data, one can also compute the range of Treatment values that correspond
to a single Control group value and return that range (See argument 'discrete.range'.)
This range is typically wider than the confidence interval of the BQTE estimate
as BQTE is defined as the expected value over those Treatment group values that
correspond to the quantile levels of a particular Control roup value.

Can also compute upper tail BQTE (UTBQTE) and lower tail BQTE (LTBQTE) at each point in 'at'.
UTBQTE, at outcome value 'x' is the difference in the mean outcome values
 of the highest 1-F(x) of the two populations, where F is the CDF of the Control population.
LTBQTE, at outcome value 'x' is the difference in the mean outcome values
 of the lowest F(x) of the two populations, where F is the CDF of Control population.
UTBQTE(x) is an upper bound for the average treatment effect of that part of
the Control population who have outcome value >= x.
LTBQTE(x) is a lower bound for the average treatment effect of that part of
the Control population who have outcome value <= x.
These bounds are general in the sense that they do not require an assumption
about how the treatment operates, for example, they do not assume anything about
whether the treatment preserves order.
Note that the outcome value 'x' is included in the definition of both UTBQTE(x)
and in LTBQTE(x).
The confidence intervals of UTBQTE and LTBQTE are estimated using bootstrap
and the point estimates are computed either using bagging (if bagging = TRUE)
or without bagging (if bagging = FALSE).
}
\examples{
set.seed(12); Tr = rweibull(100,1.2,1); Co = rnorm(100,3,5)
bqte(Tr, Co)
}
