% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{bqte_doksum}
\alias{bqte_doksum}
\title{Doksum estimator}
\usage{
bqte_doksum(at, Treatment, Control)
}
\arguments{
\item{at}{Control values where estimation is done.}

\item{Treatment}{Observed values of the treatment group}

\item{Control}{Observed values of the control group}
}
\value{
vector of estimates at the values given in 'at'.
}
\description{
Estimator of BQTE published by Doksum (1974), Annals of Statistics Vol.2 No.2 267-277.
}
\examples{
set.seed(12); Tr = rweibull(30,1.2,1); Co = rnorm(30,3,5)
bqte_doksum(at = c(3,5), Tr, Co)
}
