% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{exact_bqte_discrete}
\alias{exact_bqte_discrete}
\title{BQTE for discrete distribution}
\usage{
exact_bqte_discrete(at, x, pr.con, pr.trt, verbose = TRUE)
}
\arguments{
\item{at}{value at which bqte is computed.}

\item{x}{outcome values corresponding to probabilities in
pr.con and pr.trt. Value 'at' must be found among 'x'.}

\item{pr.con}{(unnormalized) probability mass function of the control group.}

\item{pr.trt}{(unnormalized) probability mass function of the treatment group.}

\item{verbose}{(default TRUE), if TRUE, prints info on the console.}
}
\value{
BQTE at value 'at'
}
\description{
BQTE for discrete distribution
}
\examples{
exact_bqte_discrete(3, x = 1:4, pr.con = c(0.1,0.2,0.3,0.4), pr.trt = c(0.3,0.5,0.15,0.05))
}
