% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{plot_bqte}
\alias{plot_bqte}
\title{Plots BQTE}
\usage{
plot_bqte(
  x,
  plot.ci = TRUE,
  plot.rbqte = FALSE,
  plot.range = FALSE,
  pch = 19,
  col = "black",
  col.ci = "red",
  col.range = "gray",
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  xaxs = "r",
  yaxs = "r",
  xaxp = NULL,
  yaxp = NULL,
  cex = 1,
  cex.axis = 1,
  cex.lab = 1,
  main = "",
  lwd = 1
)
}
\arguments{
\item{x}{data.frame returned by function bqte( )}

\item{plot.ci}{if TRUE, plots confidence intervals around estimates using color col.ci}

\item{plot.rbqte}{if TRUE, plots relative BQTEs, otherwise plots actual BQTEs}

\item{plot.range}{if TRUE, plots bqte$range results (requires discrete data)}

\item{pch}{pointstyle, if NULL, then lines are drawn instead of separate points}

\item{col}{color}

\item{col.ci}{color of confidence interval}

\item{col.range}{color of range}

\item{xlim}{range of x-coordinates}

\item{ylim}{range of y-coordinates}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{xaxs}{as R's standard plotting parameter}

\item{yaxs}{as R's standard plotting parameter}

\item{xaxp}{as R's standard plotting parameter}

\item{yaxp}{as R's standard plotting parameter}

\item{cex}{as R's standard plotting parameter}

\item{cex.axis}{as R's standard plotting parameter}

\item{cex.lab}{as R's standard plotting parameter}

\item{main}{as R's standard plotting parameter}

\item{lwd}{as R's standard plotting parameter}
}
\value{
none
}
\description{
Plots backtransformed quantile treatment effect (BQTE) on direct or relative scale.
}
\examples{
set.seed(12); Tr = rweibull(30,1.2,1); Co = rnorm(30,3,5)
plot_bqte(bqte(Tr, Co))
}
