% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{plot_qte}
\alias{plot_qte}
\title{Plot quantile treatment effect}
\usage{
plot_qte(
  x,
  plot.ci = TRUE,
  plot.rqte = FALSE,
  pch = 19,
  col = "black",
  col.ci = "gray",
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  xaxs = "r",
  yaxs = "r",
  xaxp = NULL,
  yaxp = NULL,
  lwd = 1
)
}
\arguments{
\item{x}{data.frame returned by the qte () function.}

\item{plot.ci}{if TRUE, plots confidence intervals around estimates.}

\item{plot.rqte}{if TRUE, plots relative QTEs, otherwise plots actual QTEs.}

\item{pch}{pointstyle, if NULL, then lines are drawn instead of separate points}

\item{col}{color}

\item{col.ci}{color of confidence interval}

\item{xlim}{range of x-coordinates}

\item{ylim}{range of y-coordinates}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{xaxs}{as R's standard plotting parameter}

\item{yaxs}{as R's standard plotting parameter}

\item{xaxp}{as R's standard plotting parameter}

\item{yaxp}{as R's standard plotting parameter}

\item{lwd}{as R's standard plotting parameter}
}
\value{
none
}
\description{
Plot quantile treatment effect (QTE) on direct or relative scale
}
\examples{
set.seed(12); Tr = rweibull(30,1.2,1); Co = rnorm(30,3,5)
plot_qte(qte(Tr, Co))
}
