% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{plot_tbqte}
\alias{plot_tbqte}
\title{Plot tail BQTE}
\usage{
plot_tbqte(
  x,
  utbqte = TRUE,
  plot.ci = TRUE,
  plot.relative = FALSE,
  pch = NULL,
  col = "black",
  col.ci = "gray",
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  xaxs = "r",
  yaxs = "r",
  xaxp = NULL,
  yaxp = NULL,
  cex = 1,
  cex.axis = 1,
  cex.lab = 1,
  main = "",
  lwd = 1
)
}
\arguments{
\item{x}{data.frame returned by function bqte( ,tails = TRUE).}

\item{utbqte}{if TRUE, prints UTBQTE else prints LTBQTE.}

\item{plot.ci}{if TRUE, plots confidence intervals around estimates using color col.ci.}

\item{plot.relative}{if TRUE, plots relative TBQTEs, otherwise plots direct TBQTEs.}

\item{pch}{pointstyle if NULL, then lines are drawn instead of separate points.}

\item{col}{color.}

\item{col.ci}{color of confidence interval.}

\item{xlim}{range of x-coordinates.}

\item{ylim}{range of y-coordinates.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{xaxs}{as R's standard plotting parameter.}

\item{yaxs}{as R's standard plotting parameter.}

\item{xaxp}{as R's standard plotting parameter.}

\item{yaxp}{as R's standard plotting parameter.}

\item{cex}{as R's standard plotting parameter.}

\item{cex.axis}{as R's standard plotting parameter.}

\item{cex.lab}{as R's standard plotting parameter.}

\item{main}{as R's standard plotting parameter.}

\item{lwd}{as R's standard plotting parameter.}
}
\value{
none
}
\description{
Plot either lower (LTBQTE) or upper (UTBQTE) tail BQTE on direct or relative scale.
For definitions of these measures, see help of bqte().
}
\examples{
set.seed(12); Tr = rweibull(30,1.2,1); Co = rnorm(30,3,5)
plot_tbqte(bqte(Tr, Co, tails = TRUE), utbqte = FALSE, pch = 19)
}
