% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{plot_tbqte_both}
\alias{plot_tbqte_both}
\title{Plot simultaneously both tail BQTEs}
\usage{
plot_tbqte_both(
  x,
  x.utbqte = NULL,
  x.ltbqte = NULL,
  plot.ci = TRUE,
  plot.relative = FALSE,
  pch.utbqte = 25,
  pch.ltbqte = 24,
  col.utbqte = "blue",
  col.ltbqte = "purple",
  col.utbqte.ci = "blue",
  col.ltbqte.ci = "purple",
  jitter = c(0, 0),
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  xaxs = "r",
  yaxs = "r",
  xaxp = NULL,
  yaxp = NULL,
  cex = 1,
  cex.axis = 1,
  cex.lab = 1,
  main = "",
  lwd = 1
)
}
\arguments{
\item{x}{data.frame returned by function bqte( ,tails = TRUE).}

\item{x.utbqte}{control group's outcome value from which utbqte is plotted towards the upper tail.
If NULL, then the lowest point of data in 'x' is used.}

\item{x.ltbqte}{control group's outcome value from which ltbqte is plotted towards the lower tail.
If NULL, then highest point of data in 'x' is used.}

\item{plot.ci}{If TRUE, plots confidence intervals around estimates using colors.
col.utbqte.ci and col.ltbqte.ci}

\item{plot.relative}{If TRUE, plots relative TBQTEs, otherwise plots direct TBQTEs.}

\item{pch.utbqte}{pointstyle for utbqte. If NULL then the corresponding estimates are shown by a line}

\item{pch.ltbqte}{pointstyle for utbqte. If NULL then the corresponding estimates are shown by a line}

\item{col.utbqte}{color of utbqte.}

\item{col.ltbqte}{color of ltbqte.}

\item{col.utbqte.ci}{color of confidence interval for utbqte.}

\item{col.ltbqte.ci}{color of confidence interval for ltbqte.}

\item{jitter}{vector of length 2, gives the horizontal offsets for points to avoid overlaps.
between plotted LTBQTE and UTBQTE estimates and intervals.
1st value is offset for UTBQTE, 2nd value is offset for LTBQTE
default is (0,0), i.e., no offset.}

\item{xlim}{range of x-coordinates.}

\item{ylim}{range of y-coordinates.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{xaxs}{as R's standard plotting parameter.}

\item{yaxs}{as R's standard plotting parameter.}

\item{xaxp}{as R's standard plotting parameter.}

\item{yaxp}{as R's standard plotting parameter.}

\item{cex}{as R's standard plotting parameter.}

\item{cex.axis}{as R's standard plotting parameter.}

\item{cex.lab}{as R's standard plotting parameter.}

\item{main}{as R's standard plotting parameter.}

\item{lwd}{as R's standard plotting parameter.}
}
\value{
none
}
\description{
Plot in the same figure both UTBQTE and LTBQTE on direct or relative scale.
For definitions of these measures, see help of bqte()
}
\examples{
set.seed(12); Tr = rweibull(30,1.2,1); Co = rnorm(30,3,5)
plot_tbqte_both(bqte(Tr, Co, tails = TRUE))
}
