% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{qte}
\alias{qte}
\title{Estimate quantile treatment effects.}
\usage{
qte(
  Treatment,
  Control,
  at = NULL,
  qte.conf = 0.95,
  B = 2000,
  bagging = TRUE,
  exact.sample.quantiles = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Treatment}{vector of outcome values (e.g. survival times) in Treatment group.}

\item{Control}{vector of outcome values in Control group.}

\item{at}{vector of quantiles at which QTE is estimated.
By default, between 10 and 20 points between the quantile levels
of 10/N and 1 - (10/N) where N = min(c(length(Control),length(Treatment))).}

\item{qte.conf}{(default 0.95), confidence level for QTE estimate.}

\item{B}{(default 2000), number of bootstrap samples.}

\item{bagging}{(default TRUE), if TRUE, reports bagging estimate (mean over bootstrap samples)
otherwise reports the direct estimate from the data.}

\item{exact.sample.quantiles}{(default TRUE), if TRUE, uses empirical sample quantiles (type = 1)
if FALSE, uses R's default quantiles (type = 7), where 'type' is
 argument of R's quantile() function.}

\item{verbose}{(default TRUE) if TRUE, prints parameters and recommendations on console.}
}
\value{
data frame with 7 columns

at, quantile level;

qte, quantile treatment effect;

qte.low, lower CI end point for 'qte';

qte.up, upper CI end point for 'qte';

rqte, relative quantile treatment effect with respect to control outcome value;

rqte.low, lower CI end point for 'rqte';

rqte.up, upper CI end point for 'rqte';
}
\description{
We have observed outcome values, such as survival or recovery times,
in the treatment group and in the control group.
Quantile treatment effect (QTE) is the difference between the two groups
as a function of the quantile levels.
}
\details{
Fix a grid of quantile levels where QTEs are computed (parameter 'at').
 Recommendation:
  Do not try to estimate QTEs at values that are outside
  quantile levels (10/N, 1 - (10/N)),
  where N is the minimum of sample sizes of treated and controls because of low accuracy at the tails.

Estimate the quantile treatment effects as difference between the quantiles
 of Treatment group and Control group

Use bootstrapping to estimate the uncertainty of QTE,
 by resampling with replacement a set of Treatment and Control group values that
 have the same sample size as the original Treatment and Control groups, respectively,
 and by applying the procedure explained above to each bootstrap sample.
 Note that bootstrapping accounts for uncertainty of quantiles of both groups.

The final estimate of QTE is the mean over bootstrap samples ("bagging estimate")
 and the confidence interval for QTE is estimated from the quantiles of the bootstrap sample.
NOTE: In empirical evaluation it was observed that bagging estimate tend to have smaller
      mean square error in discrete distributions than the direct estimate from the data and
     therefore the bagging estimate is used by default. This can be switched by parameter "bagging".
}
\examples{
set.seed(12); Tr = rweibull(100,1.2,1); Co = rnorm(100,3,5)
qte(Tr, Co)
}
