% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqte.R
\name{quantiles_from_discrete}
\alias{quantiles_from_discrete}
\title{Quantile from a discrete distribution}
\usage{
quantiles_from_discrete(p, pr)
}
\arguments{
\item{p}{quantile levels.}

\item{pr}{probability mass vector of non-negative values.}
}
\value{
list with two vectors: (1) quantile.level and (2) indexes.
        'quantile level' equals input vector 'p' and
        indexes contain the indexes of 'pr' corresponding to
        the quantile levels in 'p'.
}
\description{
Quantile from a discrete distribution
}
\examples{
quantiles_from_discrete(p = c(0.3,0.5), pr = c(0.2,0.2,0.3,0.3))
}
